rwlock: Fix memory order in rwlock_acquire
Since load+cmpxchg is not atomic, cmpxchg must have acquire ordering,
not the load. Must not allow loads in the critical section to be
speculated before the cmpxchg because the lock could have been acquired
and released (and the associated data structure modified) by another
thread between the load and the cmpxchg.

Change-Id: Ie53b0e6ad9255a2d14760145309d489e8b2afd82
Reviewed-on: https://eu-gerrit-1.euhpc.arm.com/167884
Tested-by: Ola Liljedahl <ola.liljedahl@arm.com>
Reviewed-by: Ola Liljedahl <ola.liljedahl@arm.com>