Add bm_hashtab hash table benchmark
Minor optimizations in hashtable and hopscotch

There is a slight change in the lookup functions of the hashtable and
hopscotch API's. The hazard pointer must be initialized by the user and
must be released by the user even if a lookup fails. This enables re-use
of an allocated hazard pointer over multiple lookup calls and should
thus decrease lookup overhead.

Change-Id: I6ef0a0dbe2355ecd7a7765010edf5bb5361a5ffe
Reviewed-on: https://eu-gerrit-1.euhpc.arm.com/c/attk/progress64/+/217971
Tested-by: Ola Liljedahl <ola.liljedahl@arm.com>
Reviewed-by: Ola Liljedahl <ola.liljedahl@arm.com>