rwlock: Strengthen orderings, document synchronizations.
Strengthen ordering when waiting for readers to leave the lock. This
removes bm_lock false positives as bm_lock readers actually write to the
protected data (in order to show their presence). Real world readers
would not write to the protected data and the previous ordering should
be enough to maintain acquire/release ordering between writers.

Add comments documenting the synchronize-with relationships.

Change-Id: I7ee2b67ec34902b86d7b19ee663bce8341aa2b5a
Reviewed-on: https://eu-gerrit-1.euhpc.arm.com/c/attk/progress64/+/499463
Tested-by: Ola Liljedahl <ola.liljedahl@arm.com>
Reviewed-by: Ola Liljedahl <ola.liljedahl@arm.com>