ringbuf: Use ICAS when acquiring slots on aarch64
On aarch64, use icas16() to read the endpoint metadata (head&tail etc).
This generally improves scalability (together with the non-blocking
release slots implemention).

The ring buffer flags are moved to the 6 least significant bits of the
ring buffer pointer. These bits are otherwise unused since the ring buffer
structure is at least 64-byte aligned. This avoids dereferencing the
ring buffer pointer to detect which mode should be used for enqueue or
dequeue.

Change-Id: Id7797b01ea3d1f75542c8a73726fe13fd6394bdf
Reviewed-on: https://eu-gerrit-1.euhpc.arm.com/c/attk/progress64/+/700775
Tested-by: Ola Liljedahl <ola.liljedahl@arm.com>
Reviewed-by: Ola Liljedahl <ola.liljedahl@arm.com>