Introduce atomics API
Introduce custom atomics API. The API is essentially equivalent to the GCC
__atomic builtins so conversion is straight-forward.
By adding a layer of abstraction, we can add additional functionality to the
atomic operations without modifying the implementation of the different
datatypes.
128-bit lockfree atomics (e.g. lockfree_compare_exchange) are also supported
by the API.
The implementations of the following datatypes have been converted to use the
new API: clhlock, lfstack, mcslock and msqueue.

Change-Id: If9055286300bd57d0eabba0a22e03f7303acfbad
Reviewed-on: https://eu-gerrit-1.euhpc.arm.com/c/attk/progress64/+/710446
Tested-by: Ola Liljedahl <ola.liljedahl@arm.com>
Reviewed-by: Ola Liljedahl <ola.liljedahl@arm.com>